+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   `Service::set_failure_actions_on_non_crash_failures`, 
   `Service::get_failure_actions_on_non_crash_failures`)
 - Add support to pause and continue services. (See: `Service::pause` and `Service::resume`)
+- Use `QueryServiceStatusEx` when querying service status. Allows getting the process ID of a
+  running service
 
 ### Changed
 - Bumped the MSRV to 1.34, because of err-derive upgrade which depend on quote, to use